% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{add_3d_coord}
\alias{add_3d_coord}
\title{Adds 3D coordinates to dataframe}
\usage{
add_3d_coord(mydf, R = 1)
}
\arguments{
\item{mydf}{Input dataframe}

\item{R}{Radius of the earth; defaults to unit circle but uses R=6361 if
"earth" is used as the argument}
}
\value{
Returns mydf with coord1, coord2, and coord3 added
}
\description{
Adds 3D coordinates to dataframe
}
