% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.R
\name{add_means}
\alias{add_means}
\title{Add means from GpGp model fit back to data}
\usage{
add_means(yeardata, model_fit)
}
\arguments{
\item{yeardata}{Data from a particular year}

\item{model_fit}{GpGp model fit containing design matrix and parameters}
}
\description{
Add means from GpGp model fit back to data
}
