% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_data.R
\name{augment_data}
\alias{augment_data}
\title{Augments data with parameter values}
\usage{
augment_data(
  pred_locs,
  myparams,
  varname_list = NULL,
  linkfuns = default_linkfuns
)
}
\arguments{
\item{pred_locs}{Locations for kriging the parameter fields}

\item{myparams}{Basis parameter values for kriging}

\item{varname_list}{List of parameter fields to krig}

\item{linkfuns}{List that contains an entry for each parameter field
specified in varname_list}
}
\value{
Returns a data frame that is a copy of the input data but with new
parameter values for each of the fields specified in varname_list
}
\description{
Adds parameters to a given data frame by kriging the basis values specified in
"myparams" to the locations in "mydata"
}
