% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veccmat_functions.R
\name{build_veccmat_list_grouped}
\alias{build_veccmat_list_grouped}
\title{Computes list of cholesky factors of the Vecchia precision matrix}
\usage{
build_veccmat_list_grouped(
  augdata_ordered,
  grouping_list,
  yearlist = 2007:2016,
  corrfun = cyl_cor_single,
  ncores = 1,
  verb = F
)
}
\arguments{
\item{augdata_ordered}{Ordered data over which to compute the factors. Data
should be 'augmented' meaning that each row has the values of the kernel
parametmers for that locations; this can be added to a dataframe by calling
"augment_data" with a parameter object}

\item{grouping_list}{The list of grouping information lists for
the process}

\item{yearlist}{The list of years for creating the factors}

\item{corrfun}{Correlation function to use, default is cyl_cor_single}

\item{ncores}{If ncores>1, parallelization is used}

\item{verb}{Depreciated}
}
\value{
Returns a list of cholesky factors of the Vecchia
precision matrix corresponding to the years in yearlist
}
\description{
Computes list of cholesky factors of the Vecchia precision matrix
}
