% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylind_simulation_functions.R
\name{chol_lik}
\alias{chol_lik}
\title{Cholesky likelihood function}
\usage{
chol_lik(z, ranges, locs, corfun)
}
\arguments{
\item{z}{Simulated observed values}

\item{ranges}{Vector of range values}

\item{locs}{Locations}

\item{corfun}{Function for computing correlations}
}
\description{
Cholesky likelihood function
}
