% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{chord_cor_double}
\alias{chord_cor_double}
\title{Helper function for computing correlations}
\usage{
chord_cor_double(ii, jj, input1, input2)
}
\arguments{
\item{ii}{First index}

\item{jj}{Second index}

\item{input1}{First location dataframe (needs coord1,coord2,coord3)}

\item{input2}{Second location dataframe (needs coord1,coord2,coord3)}
}
\description{
Helper function for computing correlations
}
