% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.R
\name{compute_cv_df}
\alias{compute_cv_df}
\title{Computes cross-validation scores and standard errors}
\usage{
compute_cv_df(
  augdata_ordered,
  cv_params,
  cv_function = cv_scores,
  cv_type = "lofo",
  yearlist = 2007:2016,
  nsamp = -1,
  ncores = 1,
  cv_options = NULL
)
}
\arguments{
\item{augdata_ordered}{The input data augmented with the kernel parameter
values at each location (can be created with function "augment_data")}

\item{cv_params}{List of parameters to the cross-validation function;
should be the same length as yearlist. If cvtype=standard this is a list of
veccmat objects, if cvtype=levitus this should be a distance matrix between
each of the observations in each year in kilometers, and if cvtype=gpgp this
should be a fitted GpGp model}

\item{cv_function}{Default is "cv_scores", an availabel option is
"levitus_preds".}

\item{cv_type}{lofo (leave-one-float-out) vs lowo (leave-one-window-out)}

\item{yearlist}{The list of years over which to compute the validation scores}

\item{nsamp}{If nsamp=-1 computes scores for all observation locations; if
not validation scores are computed for a selection of nsamp indices for
each year}

\item{ncores}{If ncores>1, how many cores should be used for parallelization}

\item{cv_options}{Options for cv; if lowo is chosen you can input the width
of the window to withold}
}
\value{
Returns a dataframe the same size of "augdata_ordered" with two
additional columns, validval and validsd, giving the mean and standard
deviation of the leave-one-float-out validation score at that point
}
\description{
Computes cross-validation scores and their associated prediction
errors using either lofo (leave-one-float-out) or lowo (leave-one-window-out)
cross validation methods
}
