% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditioning_set_functions.R
\name{compute_grouped_conditioning_sets}
\alias{compute_grouped_conditioning_sets}
\title{Computes grouped conditioning sets for use in the Vecchia approximation.}
\usage{
compute_grouped_conditioning_sets(NNarray, trivial = F)
}
\arguments{
\item{NNarray}{Matrix containing indices of nearest neighbors; generally
returned from compute_nnarray}

\item{trivial}{If true returns grouping object with no grouping done}
}
\description{
Computes grouped conditioning sets for use in the Vecchia approximation.
}
