% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditioning_set_functions.R
\name{compute_grouping_list}
\alias{compute_grouping_list}
\title{Computes list of grouped conditioning sets for each year in the given data}
\usage{
compute_grouping_list(locations, m, yearlist = 2007:2016, verb = F, ncores = 1)
}
\arguments{
\item{locations}{Ordered data on which to compute nearest neighbors}

\item{m}{Number of nearest neighbors to use in constructing the ungrouped
vecchia conditioning sets.}

\item{yearlist}{List of years over which to compute the group sets.}

\item{verb}{T/F, if T displays a progress bar}

\item{ncores}{Number of cores to use, if ncores>1 parallelization is}
}
\value{
Returns a list of length yearlist where each entry contains the
output of compute_grouped_conditioning_sets on the data for the
corresponding year.
}
\description{
Computes list of grouped conditioning sets for each year in the given data
}
