% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lincomb_predictions.R
\name{compute_lincomb_predictions}
\alias{compute_lincomb_predictions}
\title{Compute kriging distributions for linear combinations}
\usage{
compute_lincomb_predictions(
  obsdata,
  pred_locs,
  myparams,
  grouping_list_preddf,
  varname_list = default_varname_list,
  scalarvec = NULL,
  yearlist = 2007:2016,
  ncores = 1,
  ret_var = T
)
}
\arguments{
\item{obsdata}{Data containing observations}

\item{pred_locs}{Locations for predictions}

\item{myparams}{Parameter object for kriging}

\item{grouping_list_preddf}{List of group information for each year}

\item{varname_list}{List of varnames to use}

\item{scalarvec}{Vector of scalars for linear combination. If missing, this
will be inferred by assuming pred_locs is a grid}

\item{yearlist}{Years to use in kriging}

\item{ncores}{If ncores>1 parallelization is used}

\item{ret_var}{T/F, if T returns the kriging variance of the linear combination
as well as the mean}
}
\value{
Returns the value of pred%*%scalarvec, where pred is the vector
of kriging predictions located at pred_locs
}
\description{
Computes posterior distribution for a linear combination of
predicted values
}
