% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditioning_set_functions.R
\name{compute_nn_clusters}
\alias{compute_nn_clusters}
\title{Finds clusters of sets of nearest neighbors for grouping sets. Function
used internally and is adapted from GpGP package (see in-function comments
for citation details).}
\usage{
compute_nn_clusters(NNarray)
}
\arguments{
\item{NNarray}{Matrix continaing indices of nearest neighbors; generally
returned from compute_nnarray}
}
\value{
Returns a list of clustered arrays of nearest neighbors of length
m, where m is the second dimension of NNarray
}
\description{
Finds clusters of sets of nearest neighbors for grouping sets. Function
used internally and is adapted from GpGP package (see in-function comments
for citation details).
}
