% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditioning_set_functions.R
\name{compute_nnarray}
\alias{compute_nnarray}
\title{Computes array of nearest neighbors to be used for Vecchia process
conditioning sets}
\usage{
compute_nnarray(locations, m, verb = F, ncores = 1)
}
\arguments{
\item{locations}{Ordered data on which to compute nearest neighbors}

\item{m}{Number of nearest neighbors}

\item{verb}{T/F, if T displays a progress bar}

\item{ncores}{Number of cores to use, if ncores>1 parallelization is
used}
}
\value{
Returns a matrix of size (nlocs,m) where nlocs is the number of
locations specified in "ordered_input".
}
\description{
Computes array of nearest neighbors to be used for Vecchia process
conditioning sets
}
