% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.R
\name{convert_gpgp_fit_to_cvparams}
\alias{convert_gpgp_fit_to_cvparams}
\title{Convert GpGp model fit object to veccmat object}
\usage{
convert_gpgp_fit_to_cvparams(
  gpgp_fit_list,
  ordered_data,
  corrfun,
  yearlist = 2007:2016
)
}
\arguments{
\item{gpgp_fit_list}{List of model fits GpGp::fit_model}

\item{ordered_data}{Observation data}

\item{corrfun}{Correlation function to use}

\item{yearlist}{Lits of years to use}
}
\value{
Returns a list of "veccmat" objects
}
\description{
Convert GpGp model fit object to veccmat object
}
