% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_nsgp_to_cvparams.R
\name{convert_nsgp_to_cvparams}
\alias{convert_nsgp_to_cvparams}
\title{Function to convert BayesNSGP samples to cv_params object}
\usage{
convert_nsgp_to_cvparams(
  nsgp_vals,
  knot_points,
  argo_data_chord,
  grouping_list
)
}
\arguments{
\item{nsgp_vals}{Samples obtained from fitting a BayesNSGP model}

\item{knot_points}{Knot points used in model fit}

\item{argo_data_chord}{Argo data with 3D euclidean coordinates}

\item{grouping_list}{List of groups to use}
}
\description{
Function to convert BayesNSGP samples to cv_params object
}
