% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_params_to_statiso.R
\name{convert_params_to_statiso}
\alias{convert_params_to_statiso}
\title{Convert nonstationary anisotropic parameters to stationary or isotropic}
\usage{
convert_params_to_statiso(myparams, myaugdata, mymodel)
}
\arguments{
\item{myparams}{Parameters to convert}

\item{myaugdata}{Augmented data to use in the conversion}

\item{mymodel}{Contains list of parameters to convert to stationarity and a
boolean "iso" for whether or not the model should be converted to isotropic}
}
\description{
Convert nonstationary anisotropic parameters to stationary or isotropic
}
