% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{convert_theta_lat_to_effective_range_deg}
\alias{convert_theta_lat_to_effective_range_deg}
\title{Converts theta lat to effective range in degrees}
\usage{
convert_theta_lat_to_effective_range_deg(thetalat)
}
\arguments{
\item{thetalat}{Theta lat}
}
\description{
Converts theta lat to effective range in degrees
}
