% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{convert_theta_lon_to_effective_range_deg}
\alias{convert_theta_lon_to_effective_range_deg}
\title{Converts theta lon to effective range in degrees}
\usage{
convert_theta_lon_to_effective_range_deg(thetalon)
}
\arguments{
\item{thetalon}{Theta lon}
}
\description{
Converts theta lon to effective range in degrees
}
