% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veccmat_functions.R
\name{create_veccmat_grouped}
\alias{create_veccmat_grouped}
\title{Creat grouped veccmat objects}
\usage{
create_veccmat_grouped(
  augdata_ordered,
  groupobj,
  ncores = 1,
  corrfun = cyl_cor_single,
  covparms = NULL
)
}
\arguments{
\item{augdata_ordered}{Ordered data over which to compute the factors. Data
should be 'augmented' meaning that each row has the values of the kernel
parametmers for that locations; this can be added to a dataframe by calling
"augment_data" with a parameter object}

\item{groupobj}{The grouping information for the Vecchia process}

\item{ncores}{Number of cores to use, if ncores>1 then parallelization}

\item{corrfun}{Which function to use for correlations; default cylindrical
is used}

\item{covparms}{Optional, only needed if stationary corrfun is used. For other
correlation functions the non-stationary parameters for the kernel convolutions
should be included the augdata_ordered dataframe}
}
\description{
Computes cholesky factors of the Vecchia precision matrix corresponding
to a specified grouping of conditioning sets. Code inspired by similar code in
GPVecchia package, see source code for citation.
}
