% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.R
\name{cv_scores}
\alias{cv_scores}
\title{Compute cross-validation scores with veccmat object}
\usage{
cv_scores(cv_indices, obspred_df, veccmat)
}
\arguments{
\item{cv_indices}{Indices to predict in cross-validation}

\item{obspred_df}{Dataframe containing both observation and prediction
locations}

\item{veccmat}{Veccmat object corresponding to data in obspred_df}
}
\description{
Compute cross-validation scores with veccmat object
}
