% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{cyl_cor_double}
\alias{cyl_cor_double}
\title{Helper function for computing correlations}
\usage{
cyl_cor_double(ii, jj, input1, input2)
}
\arguments{
\item{ii}{First index}

\item{jj}{Second index}

\item{input1}{First dataframe}

\item{input2}{Second dataframe}
}
\description{
Helper function for computing correlations
}
