% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{cyl_cor_exact_single}
\alias{cyl_cor_exact_single}
\title{Helper function for computing correlations}
\usage{
cyl_cor_exact_single(ii, jj, myinput)
}
\arguments{
\item{ii}{First index}

\item{jj}{Second index}

\item{myinput}{First dataframe}
}
\description{
Helper function for computing correlations
}
