% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylind_simulation_functions.R
\name{cylind_approx_simulation}
\alias{cylind_approx_simulation}
\title{Runs simulation comparing exact cylindrical convolutions with
Gaussian approximation}
\usage{
cylind_approx_simulation(true_range_seq, rangefun, nrep, locs)
}
\arguments{
\item{true_range_seq}{Sequence of true range parameters}

\item{rangefun}{Function for computing non-stationary range parameters}

\item{nrep}{Number of repetitions to simulate data}

\item{locs}{Locations on which to generate data}
}
\description{
Runs simulation comparing exact cylindrical convolutions with
Gaussian approximation
}
