% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{cylindrical_correlation_exact}
\alias{cylindrical_correlation_exact}
\title{Computes cylindrical correlations using kernel convolutions}
\usage{
cylindrical_correlation_exact(loc1, loc2, range1, range2)
}
\arguments{
\item{loc1}{The first location on the longitudinal circle}

\item{loc2}{The second location on the longitudinal circle}

\item{range1}{The longitudinal range for the first location}

\item{range2}{The longitudinal range for the second location}
}
\value{
Returns the value of the exact cylndrical correlation; see supplementary
material for details
}
\description{
Computes the exact longitudinal correlation kernel convolution between
locations loc1 and loc2 with longitudinal ranges range1 and range2
respectively. Since this is over a circular domain Gaussian error function
calls are required.
}
