% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{cylindrical_correlation_gaussian}
\alias{cylindrical_correlation_gaussian}
\title{Cylindrical correlations using Gaussian approximation}
\usage{
cylindrical_correlation_gaussian(loc1, loc2, range1, range2)
}
\arguments{
\item{loc1}{The first location on the longitudinal circle}

\item{loc2}{The second location on the longitudinal circle}

\item{range1}{The longitudinal range for the first location}

\item{range2}{The longitudinal range for the second location
Returns the Gaussian approximation to the full convolution, which is accurate
if the effective range is less than about 100 degrees; see supplementary
material for details.}
}
\description{
Computes the approximate longitudinal correlation kernel convolution between
locations loc1 and loc2 with longitudinal ranges range1 and range2
respectively.
}
