% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{euclidean_correlation_convolution}
\alias{euclidean_correlation_convolution}
\title{Computes euclidean correlations using kernel convolutions}
\usage{
euclidean_correlation_convolution(loc1, loc2, range1, range2)
}
\arguments{
\item{loc1}{The first location on the longitudinal circle}

\item{loc2}{The second location on the longitudinal circle}

\item{range1}{The longitudinal range for the first location}

\item{range2}{The longitudinal range for the second location}
}
\value{
Returns the exact value of of the convolution in the Euclidean (latitudinal)
dimension
}
\description{
Computes the exact latitudinal correlation kernel convolution between
locations loc1 and loc2 with latitudinal ranges range1 and range2 respectively
}
