% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{gaussian_integral}
\alias{gaussian_integral}
\title{Internal function for computing
exact kernel convolutions over the cylindrical dimension.}
\usage{
gaussian_integral(loc1, loc2, range1, range2, upper, lower)
}
\arguments{
\item{loc1}{First location}

\item{loc2}{Second location}

\item{range1}{Range parameter for first location}

\item{range2}{Range parameter for second location}

\item{upper}{Upper bound}

\item{lower}{Lower bound}
}
\description{
Internal function for computing
exact kernel convolutions over the cylindrical dimension.
}
