% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{gen_deltas_from_grid}
\alias{gen_deltas_from_grid}
\title{Generates delta values on a grid for computing numerical integrals}
\usage{
gen_deltas_from_grid(grid_locs, res = NULL)
}
\arguments{
\item{grid_locs}{Grid locations}

\item{res}{Resolution of the grid; if not included will be inferred}
}
\description{
Generates delta values on a grid for computing numerical integrals
}
