% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_masked_grid.R
\name{gen_masked_grid}
\alias{gen_masked_grid}
\title{Generates grid over the sphere using the mask saved in MCMC_Input/mask.RData
(for our purposes this mask is obtained from the Roemmich-Gilson climatology)}
\usage{
gen_masked_grid(latres = 1, lonres = 1)
}
\arguments{
\item{latres}{The latitudinal resolution of the desired grid}

\item{lonres}{The longitudinal resolution of the desired grid}
}
\value{
Returns a dataframe with columns "lat_rad", "lon_rad",
"lat_degrees", and "lon_degrees" corresponding to the gridpoint
locations
}
\description{
Generates grid over the sphere using the mask saved in MCMC_Input/mask.RData
(for our purposes this mask is obtained from the Roemmich-Gilson climatology)
}
