% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanfield_sampling_funs.R
\name{get_mu_integrated_posterior}
\alias{get_mu_integrated_posterior}
\title{Computes mean/trend integrated posteriors}
\usage{
get_mu_integrated_posterior(sampleout, myparams, pred_locs)
}
\arguments{
\item{sampleout}{The output of sample_meanfield}

\item{myparams}{The current iteration's parameter object}

\item{pred_locs}{Grid for computing integrated field}
}
\value{
Returns a list containing mu0 (the posterior mean of the
integrated mean field), mu0_var (the posterior variance of the
integrated mean field), slope and slope_var analogously, and
corr_term which gives the correlation between the integrated
mean and slopes
}
\description{
Computes the mean and variance for the posterior distribution of
the globally integrated mean and trend fields as well as their
correlation
}
