% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_functions.R
\name{get_region_params}
\alias{get_region_params}
\title{For basis parameter, restricts the knot locations to a specific region and
re-krigs the parameter values.}
\usage{
get_region_params(
  inputparams,
  region_name,
  varname_list = default_varname_list,
  invlinkfuns = default_invlinkfuns
)
}
\arguments{
\item{inputparams}{Basis function parameters}

\item{region_name}{Region name, see documentation for "in_region" for
available options}

\item{varname_list}{List of names of parameters to convert}

\item{invlinkfuns}{Inverse link functions for parameter fields}
}
\value{
Returns a list of parameters with the same hyperparameters
as inputparameters
}
\description{
For basis parameter, restricts the knot locations to a specific region and
re-krigs the parameter values.
}
