% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_config_functions.R
\name{get_stationary_MLEs}
\alias{get_stationary_MLEs}
\title{Find MLEs for stationary process}
\usage{
get_stationary_MLEs(
  mydata,
  yearlist = 2007:2016,
  verb = F,
  prior_list = NULL,
  est_nugget = F,
  est_mu = T,
  est_slope = T,
  iso = F
)
}
\arguments{
\item{mydata}{dataframe with column z containing osbervations}

\item{yearlist}{list of years to use, default is 2007:2016}

\item{verb}{If T use trace=6 in optim}

\item{prior_list}{The list of coefficients for normal priors
on the log variables (if NULL or missing no priors are used)}

\item{est_nugget}{T/F if T nugget is estimated}

\item{est_mu}{T/F, if T constant mean is estimated}

\item{est_slope}{T/F, if T linear slope is estimated}

\item{iso}{T/F, if T an isotropic model will be used}
}
\value{
Returns a list containing the maximum likelihood estimates for
a stationary cylindrical model on the given data. Output contains
"theta_lat", "theta_lon", "nugget", "mu0", "phi", "slope", "likelihood",
"data_mean" and "data_var"
}
\description{
Computes maximum likelihood estimates for an isotropic cylindrical model
}
