% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_config_functions.R
\name{gp_profiled_stationary_likelihood}
\alias{gp_profiled_stationary_likelihood}
\title{Computes profiled log-likelihood for stationary Gaussian process}
\usage{
gp_profiled_stationary_likelihood(
  mydata,
  logparams,
  yearlist = 2007:2016,
  return_profiled = F,
  prior_list = NULL,
  est_mu = F,
  est_slope = F,
  iso = F
)
}
\arguments{
\item{mydata}{dataframe with column z containing osbervations}

\item{logparams}{vector of log parameters (theta_lat,theta_lon,nugget)}

\item{yearlist}{list of years to use, default is 2007:2016}

\item{return_profiled}{T/F, if T profiled parameters are returned}

\item{prior_list}{The list of coefficients for normal priors
on the log variables (if NULL or missing no priors are used)}

\item{est_mu}{T/F, if T constant mean is estimated}

\item{est_slope}{T/F, if T linear slope is estimated}

\item{iso}{T/F, if T an isotropic model will be used}
}
\value{
If return_profiled=F, returns the log-likelihood. Otherwise,
returns a list containing the log-likelihood and the profiled mu0, slope,
and phi (variance) values
}
\description{
Compute Gaussian process log-likelihood with options to profile
over the mean and variance parameters
}
