% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditioning_set_functions.R
\name{groupinfo_full_chol}
\alias{groupinfo_full_chol}
\title{Computes grouping object for full Cholesky}
\usage{
groupinfo_full_chol(nobs)
}
\arguments{
\item{nobs}{Number of observations}
}
\description{
Computes grouping object for full Cholesky
}
