% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_functions.R
\name{in_region}
\alias{in_region}
\title{Restricts the dataframe given to a specified region.}
\usage{
in_region(df, region_name)
}
\arguments{
\item{df}{Dataframe containing locations in columns "lon_degrees" and
"lat_degrees".}

\item{region_name}{Options are North Atlantic (AtlN), Pacific Northeast (PacNE),
Pacific Northwest (PacNW), Tropical Atlantic (AtlTrop), Western Tropical
Pacific (PacTropW), Eastern Tropical Pacific (PacTropE), Indian (Ind),
Southern Pacific (PacSo), Southern Indian (IndSo), Southern Atlantic
(AtlSo), and globe. See manuscript or source code for region
definitions}
}
\value{
Returns a subset of df with locations constrained to the
specified region.
}
\description{
Restricts the dataframe given to a specified region.
}
