% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_argo_field.R
\name{krig_argo_field}
\alias{krig_argo_field}
\title{Computes kriging values and optionally variances}
\usage{
krig_argo_field(input_data, pred_indices, ret_var = F)
}
\arguments{
\item{input_data}{Data containing both observation locations and
prediction locations}

\item{pred_indices}{Indices of locations in the input_data object
on which predictions should be computed; it is implied that indicies
not in this vector are observation locations.}

\item{ret_var}{T/F, if T compute and return the kriging variane}
}
\value{
Returns a list with entries "val" containing the kriged
values at the prediction locations, and "var" containing the kriging
variances (only if ret_var=T)
}
\description{
Computes kriging values and optionally variances
}
