% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_argo_field.R
\name{krig_basis_to_field}
\alias{krig_basis_to_field}
\title{Creates parameter field from basis values}
\usage{
krig_basis_to_field(
  myparams,
  predlocs,
  varname,
  linkfun,
  corrfun = cyl_cor_double
)
}
\arguments{
\item{myparams}{List containing knot locations and basis values}

\item{predlocs}{Data frame containing locations for kriging}

\item{varname}{Name of parameter field to krig}

\item{linkfun}{Link function for field}

\item{corrfun}{Correlation function to use}
}
\value{
Returns a vector of parameters at the locations specificed
by predlocs
}
\description{
Creates parameter field from basis values
}
