% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.R
\name{levitus_preds}
\alias{levitus_preds}
\title{Compute predictions using method described by Levitus et al. (2012),
see supplementary materials for details}
\usage{
levitus_preds(
  pred_indices,
  obspred_df,
  distmat_km = NULL,
  distfun = NULL,
  verb = F
)
}
\arguments{
\item{pred_indices}{Indices to predict}

\item{obspred_df}{Dataframe containing observation and prediction locations}

\item{distmat_km}{Optional, can make it faster if doing cross validation
than having to re-compute distances}

\item{distfun}{If distmat not supplied, function for computing distances;
should be geodist for great circle distances}

\item{verb}{Should progress be displayed?}
}
\value{
Dataframe with "validval" (predictions) and "validsd"
(standard errors)
}
\description{
Compute predictions using method described by Levitus et al. (2012),
see supplementary materials for details
}
