% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{mygc_dist}
\alias{mygc_dist}
\title{Compute great circle distance in radians}
\usage{
mygc_dist(l1, l2)
}
\arguments{
\item{l1}{Argument 1}

\item{l2}{Argument 2}
}
\description{
Compute great circle distance in radians
}
