% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{mygc_dist_degrees}
\alias{mygc_dist_degrees}
\title{Compute great circle distance in degrees}
\usage{
mygc_dist_degrees(l1, l2)
}
\arguments{
\item{l1}{Argument 1}

\item{l2}{Argument 2}
}
\description{
Compute great circle distance in degrees
}
