% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_functions.R
\name{order_yeardata}
\alias{order_yeardata}
\title{Orders data for each year}
\usage{
order_yeardata(unordered_data, yearlist = 2007:2016)
}
\arguments{
\item{unordered_data}{Un-ordered data to order. Must contain a "years"
column. Data for each year will be ordered separately.}

\item{yearlist}{List of years to order.}
}
\value{
Returns a dataframe the same size as "unordered_data"
}
\description{
Orders data for each year in "yearlist" according to the max-min distance
ordering
}
