% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R
\name{paramgrid_to_plots}
\alias{paramgrid_to_plots}
\title{Creates plots for each variable location on a grid}
\usage{
paramgrid_to_plots(
  plotgrid,
  configtype = "",
  breaklist,
  which_plots = c("stdev", "nugget", "efflat", "efflon", "mu", "slope")
)
}
\arguments{
\item{plotgrid}{Gridded datframe containing values to plot}

\item{configtype}{Configuration type for labels}

\item{breaklist}{List of breakpoints for each variable}

\item{which_plots}{Which variables to plot}
}
\description{
Creates plots for each variable location on a grid
}
