% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R
\name{plot_cut_diverging}
\alias{plot_cut_diverging}
\title{Plots map of observations "cut" by supplied breaks with diverging
colorbar}
\usage{
plot_cut_diverging(
  mydf,
  mybreaks,
  varsym,
  varname,
  legendname,
  mytitle = "",
  is_grid = T,
  rounddigit = 1,
  mycex = 1,
  mylabels = NULL
)
}
\arguments{
\item{mydf}{Dataframe to plot}

\item{mybreaks}{Breakpoints to use}

\item{varsym}{Symbol of variablename for legend}

\item{varname}{Name of variable to plot}

\item{legendname}{Name of legend}

\item{mytitle}{Plot title}

\item{is_grid}{Is data located on grid?}

\item{rounddigit}{How many digits to round?}

\item{mycex}{Cex value for plotting if is_grid=F}

\item{mylabels}{Labels to use in legend}
}
\description{
Plots map of observations "cut" by supplied breaks with diverging
colorbar
}
