% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{pred_with_veccmat}
\alias{pred_with_veccmat}
\title{Predictions using veccmat input}
\usage{
pred_with_veccmat(
  augdata_ordered_full,
  pred_indices,
  veccmat,
  obs_indices = !pred_indices,
  ret_var = F,
  scalarvec = NA
)
}
\arguments{
\item{augdata_ordered_full}{Data frame containing both observations and locations
as well as prediction locations}

\item{pred_indices}{Specifies which indices of augdata_ord_full are to be predicted}

\item{veccmat}{The cholesky factor of the precision matrix for the Vecchia
process to be predicted}

\item{obs_indices}{The indices in augdata_ordered_full that correspond to
observations. By default this is the complement of the indices in pred_indices}

\item{ret_var}{T/F, it T calculates and returns the kriging variance}

\item{scalarvec}{If specified will compute predictions and variances for the
linear combination scalarvec*preds}
}
\value{
If ret_var=F, returns the vector of predicted values at the specified
prediction locations. Otherwise returns a list with entries "pred" and
"var". If scalarvec is specified returns the prediction and variance for the
linear combination.
}
\description{
Computes predictions using a cholesky precision matrix for a
vecchia process. Inspired by similar code in GPVecchia package; see source
code for citation.
}
