% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_status.R
\name{record_status}
\alias{record_status}
\title{Records status of mcmc sampler}
\usage{
record_status(
  status_filename,
  iteration,
  current_likelihood,
  current_prior,
  accepts,
  variance_scaling_factor,
  start_time
)
}
\arguments{
\item{status_filename}{Location to write the sampler status.}

\item{iteration}{The current iteration.}

\item{current_likelihood}{The current likelihood.}

\item{current_prior}{The current prior value.}

\item{accepts}{The current vector of Metropolis-Hastings acceptances.}

\item{variance_scaling_factor}{The current vector of scaled proposal
variances.}

\item{start_time}{The original start time for the sampler}
}
\value{
Returns a data frame that is a copy of the input data but with new
parameter values for each of the fields specified in varname_list
}
\description{
Records the status of the mcmc sampler through appending to the file
specified in "status_filename".
}
