% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_meanfields.R
\name{resample_slopes}
\alias{resample_slopes}
\title{Re-samples slopes from marginal posterior}
\usage{
resample_slopes(mu_sampleout, nresamp, krig_grid, scalarvec = NULL)
}
\arguments{
\item{mu_sampleout}{List containing parameters and posterior covariances
for the mean and trend fields. Should be the output of "sample_mean_trend"
run with return_full_posterior=T.}

\item{nresamp}{Number of re-sampled fields to return.}

\item{krig_grid}{Grid for computing integrated mean/trend values.}

\item{scalarvec}{Values for numerical integrated; if missing will infer
from grid.}
}
\value{
Returns a list with "trend_samples", a nresample(x)ngrid matrix
where each row contains the trend values of a resampled field at the
corresponding grid-points, "intslope_samples" which is a vector of length
nresamp giving the integrated values of "trend_samples", "intmean" which
gives the posterior mean of the integrated trend, and "intvar" which
gives the posterior variance of the integrated trend.
}
\description{
Re-samples slopes from the marginal posterior conditional on
the values of the other parameters.
}
