% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcmc_sampler.R
\name{run_mcmc_sampler}
\alias{run_mcmc_sampler}
\title{Runs MCMC sampler for the cylindrical model with linear trend term}
\usage{
run_mcmc_sampler(
  ordered_data,
  initparams,
  M = 20000,
  m = "chol",
  yearlist = 2007:2016,
  pred_locs = NULL,
  grouping_list = NULL,
  grouping_list_preddf = NULL,
  outdir = "../MCMC_Output/",
  varname_list = NULL,
  corrfun = cyl_cor_single,
  var_sample_order = NULL,
  stationary_varnames = NULL,
  region_name = "globe",
  continue_from_previous = F,
  run_label = "mcmc",
  linkfuns = default_linkfuns,
  ret_last_veccmat = F,
  invlinkfuns = default_invlinkfuns,
  pred_iter = Inf,
  save_iter = Inf,
  plot_iter = Inf,
  ncores = 1
)
}
\arguments{
\item{ordered_data}{Data for running the sampler; should be already ordered.}

\item{initparams}{Parameters for initializing the sampler.}

\item{M}{number of iterations to run the sampler.}

\item{m}{Can either be the string "chol" indicating that the full
Cholesky should be used, or an integer indicating the number of Vecchia
neighbors that should be used.}

\item{yearlist}{The list of years to be used.}

\item{pred_locs}{Locations to compute predictions}

\item{grouping_list}{List of observation location groupings for Vecchia
approximation, if not included the list will be computed using the value of m.}

\item{grouping_list_preddf}{List of groupings of prediction locations
(as specified in pred_locs); if empty this will be computed in the function}

\item{outdir}{Path to directory for storing output.}

\item{varname_list}{List of variable names to sample.}

\item{corrfun}{Correlation fun to use, defaults to cylindrical}

\item{var_sample_order}{Order of variables names for sampling. Note that mu0
and slope should be specified as "mu" in this list as they are sampled together
from their joint marginal posterior distribution}

\item{stationary_varnames}{List containing which of the specified variables
should be maintained as stationary.}

\item{region_name}{Name of the region to restrict to; see documentation
of "in_region" for available  options. Data does not need to be restricted
to the region beforehand.}

\item{continue_from_previous}{T/F, if T continue from previous iteration.}

\item{run_label}{Label used in saving the output. If label is reused and
continue_from_previous=F then the previous run will be overwritten}

\item{linkfuns}{List of link functions for the variables specified
in varname_list.}

\item{ret_last_veccmat}{T/F, if T returns the last veccmat list}

\item{invlinkfuns}{Inverse link functions for the variables specified
in varname list.}

\item{pred_iter}{If do_predictions=T the sampler will compute predictions
for every pred_iter iterations}

\item{save_iter}{Every save_iter iterations, the samples since the last
checkpoint will be saved in the output directory. Default is to not save
samples and rather return all of the samples at the end.}

\item{plot_iter}{At each plot_iter iterations figures showing the
parameter field maps will be saved in the output directory (note: plotting
not currently implemented)}

\item{ncores}{Number of cores to be used; if ncores>1 then parallelization
will be used}
}
\value{
Returns the last M-save_iter samples, or all M samples if
save_iter>M. If save_iter<M, earlier samples are stored every save_iter
iterations in the directory yspecified by outdir
}
\description{
Runs MCMC sampler for the cylindrical model with linear trend term
}
