% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanfield_sampling_funs.R
\name{sample_mean_trend}
\alias{sample_mean_trend}
\title{Sample mean and trend fields}
\usage{
sample_mean_trend(
  augdata_ordered,
  myparams,
  veccmat_list,
  yearlist = c(2007:2016),
  return_posterior_mean = T,
  return_full_posterior = F,
  ncores = 1,
  seed = "rand"
)
}
\arguments{
\item{augdata_ordered}{The data from which to compute the posterior.}

\item{myparams}{The basis values for the other parameter fields;
also needs to contain hyperpriors for the mean and trend fields.}

\item{veccmat_list}{List of cholesky factors of precision matrices
for each year in yearlist.}

\item{yearlist}{List of years over which to compute the posterior}

\item{return_posterior_mean}{T/F, if T returns the posterior mean, if not
returns a sample from the posterior distribution.}

\item{return_full_posterior}{T/F, if T returns the covariance matrix between
the mean and trend fields}

\item{ncores}{If ncores>1, parallelization is used}

\item{seed}{Default is 'rand' which uses a random seed, otherwise
sets a user-defined seed}
}
\value{
If return_full_posterior=F returns a parameter object with the updated
basis values for the mean and slope fields. If return_full_posterior=T, returns a
list containing the new parameters as well as the mean and covariance
matrix for the posterior distribution.
}
\description{
Samples the mean and trend parameters from their joint posterior
distribution conditioned on the other parameter fields
}
