% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R
\name{save_image}
\alias{save_image}
\title{Saves high-quality image and trims white space from result}
\usage{
save_image(image, file)
}
\arguments{
\item{image}{Image to save}

\item{file}{Filename}
}
\description{
Saves high-quality image and trims white space from result
}
