% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylind_simulation_functions.R
\name{simulate_mles}
\alias{simulate_mles}
\title{Simulates data using true parameters, and finds MLEs}
\usage{
simulate_mles(true_ranges, rangefun, locs)
}
\arguments{
\item{true_ranges}{The true ranges to use for the simulation}

\item{rangefun}{Function for computing non-stationary range parameters}

\item{locs}{Locations on which to generate data}
}
\description{
MLEs are found using both cylindrical_correlation_exact anad
 cylindrical_correlation_exact
}
