% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_likelihood_functions.R
\name{vecc_lik_from_veccmat}
\alias{vecc_lik_from_veccmat}
\title{Vecchia likelihood from veccmat object}
\usage{
vecc_lik_from_veccmat(likinput_ordered, veccmat)
}
\arguments{
\item{likinput_ordered}{The ordered data on which to evaluate the likelihood,
where observational values are stored in the z column}

\item{veccmat}{The cholesky of the precision matrix corresponding to a
vecchia process on the data}
}
\value{
Returns the log-likelihood of the observations
}
\description{
Computes the vecchia likelihood from the cholesky of the precision
matrix of a vecchia process. Based off of code with similar functionality in
the GPVecchia package, see source code for citation details.
}
