% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_likelihood_functions.R
\name{vecc_lik_over_years}
\alias{vecc_lik_over_years}
\title{Evaluates the vecchia likelihood over multiple years}
\usage{
vecc_lik_over_years(likinput_ordered, veccmat_list, yearlist = 2007:2016)
}
\arguments{
\item{likinput_ordered}{The ordered data on which to evaluate the likelihood,
where observational values are stored in the z column}

\item{veccmat_list}{List of cholesky of the precision matrices corresponding
to the years specified in yearlist}

\item{yearlist}{The years over which to calculate the likelihoods}
}
\value{
Returns a list of log-likelihoods the same length as
yearlist
}
\description{
Evaluates the vecchia likelihood over multiple years
}
